Split the Enclave object into Enclave/Thread objects
This is a more natural split since an Enclave may have multiple Threads. It
also makes it easier to port over the `enarx-keep` API.

Also, instead of asking the caller to keep track of the TCS page location, the
`Builder` keeps track of all added TCS pages and provides the list to the
`Enclave`. A new thread can be created so long as that list isn't empty. So
`Thread` objects pop a TCS page from the list during creation and push it back
during `Drop`.