Change the default value for Masked::mask
The `mask` field of `Mask` often expresses enclave requirements. Having it
default to turning all flags on means that the default is not practically
useful. Instead, it should default to the value of the `data` field, since this
field often expresses the desired features for the enclave.

The end result of this change is that the default for the `Masked` value is
that all desired features are also required. This is a somewhat useful default.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>